requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudevents==1.12.0
deprecation==2.1.0
Flask==3.1.1
functions-framework==3.8.3
google-api-core==2.25.0
google-auth==2.40.3
google-cloud-discoveryengine==0.13.9
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.72.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.3
rsa==4.9.1
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
python-dotenv==1.0.0
